refactor: presence 접속자 훅 분리 및 import 단순화
